Add deploy script
This commit adds a Ruby deploy script to simplify the process of deploying Terraform.

The script accepts the same input as the Jenkins job, and also requires GitHub credentials.

It then uses `govukcli` (in the same directory) to generate temporary AWS credentials. The GitHub and AWS credentials, along with the initial command line input and a Jenkins “crumb” are used to queue the the `Deploy_Terraform_GOVUK_AWS` job in Jenkins.

Using Jenkins allows deployments to be auditable.