Tidy up existing content publisher project
Previously we used 'content_publisher_' as a prefix to some resources
and not others; one resource also had a description. Since neither of
these has an impact on the project, this changes all resources to use
simple names and avoid the description attribute.

Previously we had the s3 writer policy in the top-level policies
directory, which made it difficult to find. This moves it to the project
directory for content publisher, noting that it's only used once.

Finally, this also removes an unused output variable.