Support passing --context to govukcli ssh
This means that the following command works for example:

  govukcli ssh --context staging-aws backend

I personally would like the govukcli ssh behaviour to be more
predictable, based on the command being run. So rather than relying on
the correct context being set, instead, passing it in every time and
being explicit. I think this is easier to understand and use, as well
as being less error prone.