Trello-1903: update prometheus lb security group
The prometheus load balancer needs to be able to forward http traffic
down to the security group attached to the prometheus instance, this
commit puts that egress rule in place.

Also removingthe instance_elb_ids from teh prometheus module, as we are
no longer building an elb at this state level, there is an application
load balancer built within infra-public-services that takes care of the
laod-balancing to prometheus.

@ronocg <conor.glynn@digital.cabinet-office.gov.uk>