Add database infrastructure for the Content Data API
The Content Performance Manager application is being renamed to the
Content Data API to better reflect it's current function.

Currently the Content Performance Manager uses the "warehouse"
infrastructure. The plan is to make the resource naming more
consistent by following the patterns for other applications like the
Publishing API.

The Terraform configuration added in this commit relates to the new
resources for the Content Data API. It's basically a copy of the
"warehouse" configuration.

Once the migration from the Content Performance Manager to the Content
Data API has been completed, the "warehouse" resources can be removed.