Set up a bucket and policy for manual elasticsearch snapshots
We're going to use snapshots for the data sync, as the current
dump/export script we use doesn't support elasticsearch 5, and it's
always better to use built-in functionality if you can.

The cluster is configured to take an automatic snapshot every day, but
these snapshots are stored in a bucket we don't have access to.  So we
need to take manual snapshots, which are written to a custom location.

*However*, terraform doesn't support configuring manual snapshots.
This has to be done by hand.  This is what the
`manual_snapshot_domain_configuration_policy` is for: it will be added
to the list of policies administrators and powerusers have, and
someone will be able to configure the elasticsearch cluster using that
policy.

The details of this all are in the AWS documentation:
https://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-managedomains-snapshots.html