Improve the Content Data API reader and writer policies
Previously I think I thought the govuk_env_sync backups were going to
a different location in the bucket, but I think I've figured out where
I actually configured them to go now.

Also, when looking at the policy in the AWS web interface, it gives a
warning something like "An Action Is Defined Without an Applicable
Resource", which I think was meaning that I needed to be more specific
about the resources in the buckets.