Parameterise elasticsearch by what manual snapshot buckets exist
Elasticsearch needs to have access to buckets across environments:

- read from the environment above (eg, pulling production data to
  staging)
- write to the current environment (eg, backing up staging)

We decided to store the bucket ARNs in govuk-aws-data to simplify
this.  There's no risk in doing this, because we only use these
buckets for the data sync.  Giving (eg) staging write access to
production doesn't let ot touch production's atcual backups, only the
ones used for the data sync.

There will be a follow-up change to govuk-aws-data.