Fix the Content Data API DB Admin reader to allow ListObjects
Through the ListBucket permission. Previously "s3:List*" was being
specified as an action, but this seems not to apply to resources that
specify files within a bucket, it must apply to the whole bucket.

This kinda makes sense, I was getting confused by the combination of
the resources and actions in the other policies.

This commit splits the actions across two statements, the first
statement applying to the bucket itself, and removes the wildcards
from the actions being specific about what's actually intended.