Add s3 bucket and iam user for content data admin
This adds an S3 bucket for content data admin to store CSVs for
users to download. The bucket has a lifecycle policy of 7 days as
the file are only temporarily made available to download. This
also creates an IAM user for content data admin to have permission
to upload to the bucket.