Remove the unnecessary ELB for the Content Data API DB Admin machine
This was copied from the app-warehouse-db-admin project, but I don't
think it's necessary there or here. I'm guessing it was just copied
from another project.

There might be a use for having an ELB if pgbouncer is running on the
DB Admin machines, as is the case in the app-db-admin project. But
hopefully pgbouncer can be removed soon, as it's risky in AWS, as
running it on the db-admin machines adds a single point of failure in
the applications connection to the database.