Module to create s3 buckets with expiration lifecycle rules
Context
We have different environments (Integration, Staging and Production) with
different lifecycle requirements yet these are created with the same code. This
means that all lifecycle rules will be the same for any buckets created with
the same code. Lifecycle rules have to be defined inline and cannot be attached
to existing buckets.

Decision
Create a module to allow for basic expiration of objects, namely expiration of
current and noncurrent objects. This commmit creates a module that allows
enable/disable of a current object expiration rule and the enable/disable of a
noncurrent object expiration rule. It also allows an override of the default 5
days for both these rules.