Start using s3 module for infra-s3-mirrors project
Context
We have a module that will improve the consistency of lifecycle rules across s3
buckets. The infra-s3-mirrors project is one of several that would benefit.

Decision
Use the module and it's defaults to ensure versioned objects are deleted for
this project. Do this while ensuring data isn't deleted and that key
infrastructure is not destroyed. This requires both code and implementation
steps.

The following approach was used to implement this across aws envs:

1. Run a plan with `./tools/build-terraform-project.sh` for the project.
2. Confirm what would be destroyed and confirm other changes.
3. Change directory to the project to run `terraform state` and rename the
entries to keep in the state file. This will also create a local backup of the
state file.  e.g. `terraform state mv old_name new_name`
4. Run a plan with `build-terraform-project.sh`, confirm expectations.
HINT: There should be `0 to destroy`.
5. Run an apply with `build-terraform-project.sh` to implement the changes.
6. Submit PR and get changes merged to master.