Add option to list all machines of a type
Sometimes it's usefull to list all the machines of a particular class,
for example when you are debugging something strange that's only
affecting a single machine.

Add a `node-list <machine>` to `ssh` to support this.

For example, running `govukcli ssh node-list whitehall_backend` should
list all the whitheall backend machines in your environment.