Remove expiration rule from elasticsearch manual snapshot bucket
Elasticsearch doesn't just look at the files in the bucket to know
what snapshots there are, it maintains a metadata file.  Expiring old
files results in this metadata file getting out of sync, which results
in a 500 error when trying to do anything other than take a new
snapshot.

There doesn't seem to be a way to regenerate the metadata based on the
real data, so automatic expiration has to go.  Perhaps we could have a
daily task on the search machines which deletes the oldest snapshot or
something.