Set the host for proxied services
Fastly uses a cache key of 'req.url, req.http.host' when deciding which
key to store a cache under. So we should set the host when proxying to
services other than the main data.gov.uk service so that they don't get
cached over data.gov.uk. This happens because we rewrite the URLs to
remove the prefix part of the URL for the proxy service. Setting the
host doesn't change the host of the backend origin service we're
connecting to, it just changes the value of the header sent to the
origin service.