Update provisioning documentation
Tfenv is the recommended way to manage Terraform versions locally, there is a
terraform-version file that has been maintained in this repo for quite a long
time now. Chtf was compatible only with Brew.

Keep the data management information in the data repository, it is more up-to-date
than what we had here.