Update infra-certificates to create all the validation DNS records
The original code is creating only one DNS validation record instead of all the
records needed to validate all the SAN entries of the certificate.

The aws_acm_certificate domain_validation_options property is a list of map
objects, we can't iterate as a normal list. It looks like:

```
validation-options = [
    {
        domain_name = myname.example.com,
        resource_record_name = _...myname.example.com.,
        resource_record_type = CNAME,
        resource_record_value = _....acm-validations.aws.
    },
    {
        domain_name = my-alt-name.example.com,
        resource_record_name = _....my-alt-name.example.com.,
        resource_record_type = CNAME,
        resource_record_value = _....acm-validations.aws.
    },
...
]
```

We are following this solution to create all the required records:
https://github.com/hashicorp/terraform/issues/17315