Configure remote state bucket region with variable
The new Training environment and the Tools environment are deployed in
a different region from the default one (eu-west-1). We need to configure
the remote state bucket location with the global variable `aws_region` to
be able to fetch the remote state file from the right location.

This change needs to be applied to the `generate-remote-state-boiler-plate.sh`
and the generated manifests to avoid being overwritten in the future.