app-puppetmaster add support for non-blued infrastructure
We want to start provisioning the Training environment app projects
without the `blue` or stack-based DNS/LB infrastructure, but the code
has to be able to support the current design in integration/staging/production
and the new one at the same time.

Already, the training doesn't have a stack Route53 zone (`infra-stack-dns-zones`
hasn't been applied), so the remote_state data resource `infra_stack_dns_zones`
can't be used in the code.

This new solution uses the `aws_route53_zone` data resource instead of the remote
state files to make it easier to understand what the code needs to implement. The
variables that determine the domain/zone name are required, which
forces the user to add the correct configuration in the data repository.