Use a lambda function to fix ES manual snapshot bucket permissions
The Elasticsearch data sync from production to staging (and staging to
integration) works by having the production Elasticsearch take a
snapshot and store it in a bucket, and then having the staging
Elasticsearch restore that snapshot.  This means that the production
and staging accounts have access to the bucket, which is owned by the
production account.

The *staging* Elasticsearch likes to put a file into the bucket called
`incompatible-snapshots`, which then prevents *production* from taking
new snapshots, as it can't read the file.

Ideally we would change the AWS managed Elasticsearch to use a
read-only role; but Elasticsearch doesn't like a snapshot repository
it can't write to.

So this PR solves the issue like so:

- Each environment gets a lambda function which listens to S3 Put
  events from the environment above (so staging listens to production
  events, integration listens to staging events).  If there is no
  environment above (as is the case for production), no lambda/S3
  permission is added, but the lambda function is still created.

- When the lambda function is notified, it checks if the new object
  is (a) not owned by the bucket owner and (b) owned by the lambda
  owner; if both of those conditions are met it grants full access to
  the owner of the bucket.

This assumes that only the bucket owner should be writing to the
bucket, but that seems a safe bet for now.

How to deploy:

1. Set the `upstream_bucket_arn` for staging and integration in
   govuk-aws-data.

2. Deploy to integration: note down the lambda ARN

3. Set the `downstream_lambda_arn` for staging in govuk-aws-data.

4. Deploy to staging: note down the lambda ARN.

5. Set the `downstream_lambd_arn` for production in govuk-aws-data.

6. Deploy to production.