Remove user checking from ES lambda code & handle all results
These are not the values we need, I tried another way (getting the
object metadata to compare the creator ID), but had a permission error
with that.  This simplified approach works for now, if there is an
error in the future which prevents the lambda from running, we'll find
out about it due to the data sync alerts in icinga.