Added infra-artefacts-bucket project to training environment
We changed the bucket policy from inline to a terraform data
object, this way it's easier to implement empty variables.
We also changed the aws_s3_access_account variable to
a list to make it more flexible, because the integration environment
needs access from more than a single aws account.

We also removed a variable that was no longer in use "deployer_role".
Finally we added the terraform backend file.

Co-authored-by: @afda16