Update app-apt for Training environment
Add backend to build app-apt in the Training environment.

Add parameters to select which domain to use with the DNS records (Training
does not use the stack domain).

This project doesn't have a second public ALB in `infra-public-services`, because
the apt service in Production is common to all the infrastructure and it hasn't
been migrated to AWS yet.