Tune the Content Data API PostgreSQL parameters
Previously there have been problems with autovacuuming for the Content
Data API database (although this was when running on Carrenza).

The following changes should potentially help with general
performance, and also the autovacuuming. The additional logging around
autovacuuming should also help refine the configuration further in the
future.

There isn't a lot of specific reasoning behind these changes, rather
they're slightly more sensible values that hopefully can be better
refined in the future.

Currently m4large instances are in use which have 8 GiB of
memory. Currently the work_mem value is 4MB, and here it's being
increased to 512MB (1/16 of the overall memory). A larger value should
permit PostgreSQL to do more in memory for large queries.

The maintainance_work_mem is currently 127MB. Due to the large tables
and indicies in the database, increasing this may help with
autovacuuming. Additionally, autovacuum_max_workers is now set to 1,
as taking it one table at a time is probably more performant.