Add ALB and app health checks to app-draft-frontend
We want to send traffic to each instance only when the application
has been deployed and is healthy. The draft-frontend class only
has an internal ELB to receive traffic. In order to use application
specific health check paths from a LB, we need to use an ALB instead
of ELB and create application specific target groups/health checks
with the lb_listener_rules module.

We initially create a new ALB, and we can move traffic to this new LB
with the `enable_alb` parameter. This parameter updates the ALIAS of
the main service record to point to the ALB instead of the original ELB.
Once we have migrated all the traffic to ALBs, we can remove the code and
resources of the original ELB.