Remove bad character at the end of the target group name
An LB target group name can't be longer than 32 characters. We are using
`format` to ensure we don't pass that limit, but sometimes this can result
in the `-` character appearing at the end of the string, which is also
not allowed an causes an error. For instance:

```
aws_lb_target_group.tg.6: Error creating LB Target Group: ValidationError:
Target group name 'draft-fron-draft-service-manual-' cannot begin or end with '-'
```

We are going to explicity remove this character if it appears at the end
of the generated string.