Create Elasticsearch 6 cluster
Trello: https://trello.com/c/4bti7M9N/744

This duplicates the deployment code we have for provisioning
the elasticsearch5 cluster for elasticsearch6.

It should work in the same way, except using elasticsearch version
6.7, the latest version available from AWS.

Once this is deployed, we will bump the number
of master nodes to 3, and change this in the
main.tf for this application.