Move code from old deprecated repo for mongodb-backup buckets
As a Reliability Engineer
I would like code for this AWS resource to exist in a current repo
So that I can make changes as expected.

These resources were initially created from a deprecated repo. To correct this
the code needed to be moved to the current repo, a new terraform state file
needed to be created with an S3 backend, resources imported and a `terraform
apply` run to confirm what is essentially a no-op.

This has been tested and completed in staging but not yet in the other
environments.

solo: jstandring-gds