Create s3 bucket with module for infra-mongodb-backup
In order to standardise lifecycle rules a module has been created. This can be
used by this project.

To make this work, the existing aws_s3_bucket.resource needs to be moved into
the module with `terraform state mv` commands. E.g. `terraform state mv
aws_s3_bucket.resource module.resource`

This has been completed in staging and production already so this code is
following up on that change.