Allow optional ASG attachment in lb_listener_rules module
At the moment, the `autoscaling_group_name` value that we pass to this
module comes from a `data` resource, that can be empty if the ASG hasn't
been created yet in an environment. This is currently happening on Training.

By adding a default empty string to this parameter, we can implement the
module even when the ASG doesn't exist. In this case, the attachment should
not be created.