Update infra-public-services listener rules forempty ASGs
In the Training environment, some of the ASGs don't exist yet. This was
causing an error when passing the parameter `autoscaling_group_name` to
the `_public_lb_rules` modules, because Terraform wasn't able to identify
the type of `names[0]` when the data resource returned no items.

The `aws_autoscaling_group` returns a single item instead of a list, which
is easier to evaluate to cover the cases when nothing is returned. We never
expect to have more than an ASG with a given name per environment, so both
data resources provide the same data.

The `rules_host_domain` in `backend_public_lb_rules` needs to remove the
environment variable in preparation for the deployment on Production, because
the Host headers are using the `publishing` domain.