Create app-related-links project for automated link generation and ingestion
This commit creates a new project `app-related-links`, which describes the required infrastructure to support automated link generation (from content store data and other sources) and link ingestion (via the Publishing API).

Policies are added for Concourse (from another account) to be able to assume role and create infrastructure in our environments, in addition to an EC2 role that Concourse can also use to determine the availability of the instances it spins up via the ASGs.

Solo: @karlbaker02