Update enable_alb implementation in frontend apps
We have applied the changes to enable the new ALBs on the frontend app
projects. The `instance_target_group_arns_length` and `instance_target_group_arns`
are actually irrelevant if the alb is enabled or disabled, because those
setting only attach the default group of the new ALB to the ASG.

Just in case we need to rollback the change, it's actually easier to
understand the plan, if the only relevant change when setting `enable_alb`
to false is a change in the DNS record ALIAS value to point to the original
ELB.

For these reasons we are going to update the code and always do the
attachment of the ALB default group.