Upgrade instance type for app-related-links generation instance
This commit upgrades the instance type used for related links generation in the `app-related-links` project from `m5.2xlarge` to `m5.4xlarge`. An `m5.4xlarge` was used during development for the testing of the related links algorithm and it was thought that a downgrade might be possible to a lesser instance type, however this has been proven false so reverting back to the original instance type.