adds waf log shipping to splunk
Adds a firehose for forwarding web acl logs to splunk and wires up the
"default" ACL to start shipping logs to splunk.

WAF ACL Logs -> Firehose -> Splunk

long term cyber will be collecting all these kinds of logs for threat
intelligence and will be doing it by exposing a cross-account firehose
destination that we could ship directly to. However at time of writing
this method is not quite ready for prime-time and so we are deploying a
firehose that will ship directly to splunk...

The cyber managed Splunk instance  will ingest the logs and make them
available both to GOV.UK members and Cyber threat intelligence.

An S3 bucket acts as a backup for the firehose stream for the case when
the splunk endpoint is unavailable.