adds lambda processor to filter waf log stream
the waf firehose stream contains entries for every single request
passing through the acl rule, which is basically every single request
through gov.uk that didn't hit the CDN. We are also shipping these
logs for all environments which are doing traffic replay so we have
compounded the volume.

Splunk costs money, we shouldn't waste money on logs that are of little
value.

We are really only interested in things that have been matched and
BLOCKED or explicitly ALLOWED by a rule, so we add a lambda to filter
out all logs records that were ALLOWED by the default rule (ie normal
perfectly fine requests).

The lambda is triggered by records passing through the firehose on the
way to splunk.