adds SQL injection match rule to default web acl
we want our web application firewall to detect and block requests
containing obvious sql injection attempts.

we add a rule to the default web acl (which is attached to the frontend
cache alb) which will de-obfuscate a set of common fields as recommended
by the aws provided cloudformation template and then match on any
detected SQL injection tokens.

initially we set this rule to ALLOW with the intension to only detect
and log requests that match... but after a few days of monitoring will
will switch to BLOCK matches of this rule.