adds XSS rule to default web acl
we want to detect and block request containing cross-site-scripting
attempts.

this adds a XSS WAF rule to the default ACL (attached to the frontend
cache ALB).

It is currently set to ALLOW all detected request so we can monitor it's
impact in production. We intend of switching this rule to BLOCK after
several days in production.