The fastly vcl `datagovuk.vcl` was static but fastly IPs do change with time. We therefore make the vcl `datagovuk.vcl` dynamic by using a template `datagovuk.vcl.tmp` which is then filled in at terraform `apply`/`plan` time.
The template `datagovuk.vcl.tmp` is filled by using a terraform external data source which runs a bash script `fastly.sh`. `fastly.sh` makes a query to fastly to get the latest fastly subnets and then uses these values to insert into the `datagovuk.vcl.tmp` to make the `datagovuk.vcl`.

In addition, there was a need to replace the condition type from `request` to `REQUEST` because of a change in fastly accepted types.