remove duplicated cloudtrail logging/monitoring
GDS enabled organisation wide cloudtrail logging/monitoring some time
ago which stores all CloudTrail events from all accounts in an immutable
bucket which are indexed by splunk and monitored by Cyber.

AWS give you the first CloudTrail in the account for free, but charge a
non-trival amount for subsequent trails.

To reduce costs we remove this old trail, and will rely on the org-wide
trails.

NOTE: we have to leave the bucket existing for now, but with the
force_destroy flag set, so that it can be cleaned up with a subsequent
terraform run