Set load balancer healthcheck paths for Licensify.
These are the same paths as currently used by the Icinga healthchecks.

`licensify-frontend` serves both the frontend UI and the `uploadlicense`
service. There is therefore one target group and one healthcheck for
`licensify-frontend`. In other words, we are treating the app instance
as the failure domain for the purpose of LB healthchecking.

`licensify-feed` is not load-balanced (because it's a single-master,
manual-failover architecture).