Make the associate-eni function more generic.
Modify the associate-eni function so that it filters out all the
in-use interfaces before selection, and allow the selection to
return more than one interface id, in that case we select the
first one.
We need this behaviour in the case of the rabbitmq cluster where
the nodes attach interface from a pool of available interfaces,
and therfore it is possible that more than one interface is
available.