Don't try to run two master instances of licensify-feed.
licensify-feed requires a single master instance. Failover to a
non-master instance requires a manual config change. In AWS, since we
have automated deployments, it is quicker to let the failed instance
be recycled by the autoscaler than to make the manual config change to
swap the master over.

For now, we therefore plan to put up with running a single instance of
licensify-feed and licensify-admin. If need be, we could do further work
to separate out licensify-admin so that it can have two instances.