Add Knowledge Graph project
This commit adds a new project `app-knowledge-graph`, which contains the required resources to spin up an instance of the Neo4j Community AMI to iterate the GOV.UK knowledge graph. We will only be deploying this project to integration, which is why this is the only environment specified. Users will be connecting over HTTPS and BOLT to query the knowledge graph, with SSH being allowed only for Karl's key-pair if it's necessary to connect to the machine (we're not provisioning via Puppet and so don't have access to all SSH keys here).