Add HTTP->HTTPS redirect at the LB for licensify-admin.
This already exists in Carrenza (on the licensify_lb nginx) and we need
to preserve the behaviour when migrating to AWS. The HTTPS redirect is
essential in rare cases, because HSTS preload cannot be relied upon
universally.

We add an ALB listener on port 80 with a default rule which serves
redirects to the same URL but with the https protocol. This way we don't
have to allow any HTTP traffic to be forwarded to the backend and we
keep the backend nginx config simple.