Add an S3 bucket to store Gatling results
This bucket can be used to store Gatling load testing results for a
longer period of time, as currently the results get wiped whenever we
need to reprovision the machine.

It's configured to allow static website access from office IP addresses
meaning we can view the results directly either in the office or by
using the VPN rather than having to download them first.