Remove unused config for separate API MongoDB cluster.
There was once a plan to have a separate API MongoDB cluster but it was
subsequently decided (after this config had been written) that the
`app-mongo` cluster would be used instead.

This config is therefore both unused and obsolete. I'm aware of no plans
to use it any time soon -- and if we should decide to use it in the
future, it's trivial to get it back.

This essentially reverts https://github.com/alphagov/govuk-aws/pull/703.