Remove unused config for separate email-alert-api RDS instance.
There was once a plan for email-alert-api to have its own RDS instance
(and a db_admin machine to go with it) but it was subsequently decided
(after this config had been written) that email-alert-api would use
the shared RDS instance instead.

This config is therefore both unused and obsolete. I'm aware of no plans
to use it any time soon -- and even if we decide to use in future then
it's trivial to get it back.

This essentially reverts https://github.com/alphagov/govuk-aws/pull/620 and
https://github.com/alphagov/govuk-aws/pull/624