Add ingress rule for Knowledge Graph SG to access PostgreSQL
This commit adds a new ingress rule for the `knowledge-graph` SG to access the PostgreSQL `primary` node (in integration) for the purposes of querying the Publishing API database directly. We run this query at present to generate a dataset of people and of roles, which we're experimenting with to improve the Knowledge Graph and the use cases we're testing off the back of it.