Fix duplicate DNS records for whitehall-admin, for real this time.
My fix in #1126 removed the wrong duplicate DNS record. We need to
create A records in `app-whitehall-backend` because this is where the
load balancer is defined. If we try to do this in
`infra-public-services`, we would introduce a cycle into the dependency
graph of terraform projects which would make things tricky to deploy.

We therefore create the A (ALIAS) records in `app-whitehall-backend` and
remove the (now unnecessary) CNAMES from `infra-public-services`. This
makes the DNS arrangements for `whitehall-admin` follow the same pattern
as those for `licensify-admin`, which is fairly well-tested at this
point.