Ensure New Mapit Nodes Are Healthy
When autoscaling, new Mapit instances are currenly not created with new EBS volumes.
This is the present state of the mapit machines:

There are two autoscaling groups each containing only one instance. If this singular instance
is terminated, a new one is launched without an EBS volume to take it’s place. Then the old EBS
volume is attached to the new node (the EBS volume contains the database for the postcode lookup).
There are two problems with the current setup:
1. The service cannot be scaled up. The current set up relies on a the single instance being
terminated so that the new instance can use the old/original volume.
2. Any new nodes that are launched are currently unhealthy and require manual intervention to
make them work. This is caused by a Postgres issue. If Postgres detects data on a volume
during installation it will not install fully ( the issue is detailed here :
https://trello.com/c/l6Yalzfd/2671-fix-mapit-deployment-automation-so-that-new-vms-come-up-healthy ).
This means the new nodes also launch with a broken Postgres installation.

This PR implements the following solution:

A new blank EBS volume is now created by the launch configuration each time a new mapit instance
is created through the autoscaling group. After Postgres is installed, puppet automatically updates
the mapit database (see related PR here:
https://github.com/alphagov/govuk-puppet/commit/8dc4a54a2783361a4efd781454ab6428a4b6606c ).

The node_group module has been modified to allow for two types of launch configuration, one for nodes
which need an EBS volume and ones for nodes that do not. The default is the launch configuration
does not create EBS volumes.

A new user data snippet has also been added to tag the mapit node EBS volumes. This is because the attach-volume user data script is no longer used. The launch configuration will automatically attach any specified volumes to the correct node on creation, but annoyingly terraform does not let you add tags to volumes created by the launch configuration.